code: the seed holds what the record spells, and two comments stop overclaiming
Four findings from the olympus review seat on PR #499, all real.

**`seed BIGINT` could not hold the record's seed.** The record spells it as
an unsigned 64-bit value and the suite's own fixtures carry values above the
signed maximum, so every run with the high bit set was unrecordable:

    INSERT INTO t(s BIGINT) VALUES (14458752852352082704);
    ERROR:  bigint out of range

The column is now `NUMERIC(20,0)`, which holds the full unsigned range
exactly and reads as the number it is. A `BIGINT` reinterpreted in two's
complement would hold it at the cost of every reader knowing that it was.
Verified: the value lands and reads back unchanged.

**The file said jsonb rejects duplicate keys. It does not.**
`'{"x":"1","x":"2"}'::jsonb` is `{"x": "2"}`: PostgreSQL keeps the last
value and drops the rest silently. The set-equality argument survives on the
half that is true, key order being normalized, and a directory holds no two
files of one name so a well-formed digest map has no duplicate to collapse.
The comment now says what jsonb does, and says that catching a malformed map
is the import's job rather than the column's.

**The version comment read as though the schema advanced the counter.** No
trigger or rule does, and none should. A clause now says the write path
reads the version, writes against it and advances it in the same
transaction, which is the same division the author comment already draws.

**The widths are chosen rather than inherited, and are now stated once.**
SQL has no unsigned integer. The seed takes NUMERIC(20,0) for the reason
above, positions and ranks and depths take INTEGER because a position is the
resident length at the draw and none approaches two billion in a session,
and token identifiers take BIGINT because the width costs nothing on a
column that is not a key.

Reapplied to PostgreSQL 18.6 from an empty database, and the seed insert
that refused before now returns the value it was given.

Per issue #454, ruling two.

+++ crates/weaver-web/migrations/0001_the_store.sql
@@ -12,6 +12,16 @@
 -- share no column any of these seven wants, so a migration between them would
 -- be a drop and a create wearing a migration's name. Git is the archive.
 
+-- The record spells several members as unsigned 64-bit values and SQL has
+-- no unsigned integer, so the widths below are chosen rather than inherited
+-- and are stated here once. The seed takes NUMERIC(20,0), which holds the
+-- full unsigned range exactly, because the records on hand carry seeds
+-- above the signed maximum. Positions, ranks and depths take INTEGER: a
+-- position is the resident length at the draw, a rank is bounded by the
+-- field election's depth, and neither approaches two billion in a session.
+-- Token identifiers take BIGINT, a vocabulary being far under it and the
+-- width costing nothing on a column that is not the table's key.
+
 -- =====================================================================
 -- The recorded half. Written only by the ingest of Spec section 3.1.
 -- No surface writes here, which is what Spec section 6's rule means.
@@ -24,7 +34,14 @@ CREATE TABLE run (
   -- The tuple. Spec 2.2: a reading without its tuple is a reading of an
   -- unnamed compound.
   record_identity  TEXT NOT NULL,
-  seed             BIGINT,
+
+  -- The record spells the seed as an unsigned 64-bit value and the suite's
+  -- own fixtures carry values above the signed maximum, so BIGINT refuses
+  -- them with `bigint out of range` and every run with the high bit set
+  -- would be unrecordable. NUMERIC(20,0) holds the full unsigned range
+  -- exactly and reads as the number it is, where a BIGINT reinterpreted in
+  -- two's complement would cost every reader the knowledge that it was.
+  seed             NUMERIC(20,0),
   sampler          JSONB NOT NULL,
   device           TEXT NOT NULL,
   compute_precision TEXT NOT NULL,
@@ -104,6 +121,12 @@ CREATE INDEX position_by_surprisal ON position (run_id, surprisal DESC);
 -- counter rather than anything the author supplies. Writes are ordered on
 -- the row rather than idempotent.
 --
+-- No trigger or rule below advances that counter, and none should. The
+-- authoring path reads the version, writes against it, and advances it by
+-- one in the same transaction, refusing where the stored version has moved.
+-- The column carries the counter and the write path carries the ordering,
+-- the same division as the author member below.
+--
 -- Spec 3.2: every authored row names its author and the member is nullable,
 -- its null meaning the store could not name an author when the row was
 -- written and never meaning the operator.
@@ -126,9 +149,16 @@ CREATE TABLE artifact (
 
   -- Spec 2.3: the weights identity is the set of per-file content digests
   -- keyed by file name. Equality is set equality and no order is imposed.
-  -- jsonb normalizes object key order and rejects duplicate keys, so a
-  -- UNIQUE on this column is set equality exactly, and this table asks for
-  -- no rolled-up digest of its own devising.
+  -- jsonb normalizes object key order, so a UNIQUE on this column is set
+  -- equality over a keyed map, and this table asks for no rolled-up digest
+  -- of its own devising.
+  --
+  -- What jsonb does NOT do is reject a duplicate key: it keeps the last
+  -- value and drops the rest silently. That does not reach this member,
+  -- because a directory holds no two files of one name, so a well-formed
+  -- digest map has no duplicate to collapse. A malformed one collapses
+  -- rather than refusing, and catching that is the import's job at Spec
+  -- 3.2 and not this column's.
   file_digests  JSONB NOT NULL UNIQUE,
 
   -- Spec 2.3: the provenance chain is recorded and is not an identity.
